Rename Novatron to Gigasaur
The _Novatron_ was renamed to the _Gigasaur_ due to a name conflict with another project. There was also a minor update to the circuit; switching from a micro to full-sized SD card. The second bit (`bit 1`) on the expansion circuit was fixed as a logic 0, but is now wired to the write-protect pin of the SD card. This bit needs to be masked low as well to make the SD card code on the Gigasaur work correctly.